shopee_image_tool/requirements.txt
streamlit
google-cloud-vision
google-generativeai
opencv-python-headless
Pillow
numpy
python-dotenv
